Refactoring of assets ☘️.
