+++ src/ignitus-Authentication/ignitus-StudentSignUp/Components/Signup.jsx
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 import isEmpty from 'lodash/isEmpty';
 import isEqual from 'lodash/isEqual';
-import loader from '../../../ignitus-Shared/ignitus-DesignSystem/ignitus-Assets/Images/ignitusLoader.gif';
+import loader from '../../../ignitus-Shared/ignitus-DesignSystem/ignitus-Assets/ignitus-Logos/ignitusLoader.gif';
 import * as t from './Constants';
 import { withErrorBoundary } from '../../../ignitus-Shared/ignitus-ErrorHandlingComponents/errorBoundary';
