+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Assets/ignitus-Logos/logo-Png/ignitusWhiteLogo.png
