+++ src/ignitus-Shared/ignitus-DesignSystem/ignitus-Organisms/ignitus-Navigation/Components/Navigation.jsx
@@ -2,8 +2,8 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { HashLink } from 'react-router-hash-link';
-import logo from '../../../ignitus-Assets/Images/ignitusBlueLogo.svg';
-import blackLogo from '../../../ignitus-Assets/Images/black-logo.svg';
+import logo from '../../../ignitus-Assets/ignitus-Logos/logo-Svg/ignitusBlueLogo.svg';
+import blackLogo from '../../../ignitus-Assets/ignitus-Logos/logo-Svg/ignitusBlackLogo.svg';
 import '../Styles/style.scss';
